Refactor FileStorage classes.
- Adds setter for FileStorage maxTransferRate.
- Adds setter for SanStorage networkLatency.
- Adds new test cases.

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/resources/HarddriveStorage.java
@@ -193,7 +193,7 @@ public class HarddriveStorage implements FileStorage {
     }
 
     @Override
-    public void setMaxTransferRate(final int maxTransferRate) {
+    public void setMaxTransferRate(final double maxTransferRate) {
         if (maxTransferRate <= 0) {
             throw new IllegalArgumentException("Max transfer rate must be greater than zero.");
         }
